extract functions to utility file
