+++ apps/client/src/common/components/input/time-input/TimeInput.tsx
@@ -1,12 +1,11 @@
-import { KeyboardEvent, useCallback, useContext, useEffect, useRef, useState } from 'react';
+import { useRef } from 'react';
 import { Button, Input, InputGroup, InputLeftElement, Tooltip } from '@chakra-ui/react';
 
 import { EventEditorSubmitActions } from '../../../../features/event-editor/EventEditor';
 import { tooltipDelayFast } from '../../../../ontimeConfig';
-import { LoggingContext } from '../../../context/LoggingContext';
-import { forgivingStringToMillis } from '../../../utils/dateConfig';
-import { stringFromMillis } from '../../../utils/time';
 import { TimeEntryField } from '../../../utils/timesManager';
+
+import useTimeInput from './timeInputUtils';
 
 import style from './TimeInput.module.scss';
 
@@ -18,119 +17,32 @@ interface TimeInputProps {
   placeholder: string;
   validationHandler: (entry: TimeEntryField, val: number) => boolean;
   previousEnd?: number;
+  value?: string;
 }
 
 export default function TimeInput(props: TimeInputProps) {
-  const { name, submitHandler, time = 0, delay = 0, placeholder, validationHandler, previousEnd = 0 } = props;
-  const { emitError } = useContext(LoggingContext);
+  const {
+    name,
+    submitHandler,
+    time = 0,
+    delay = 0,
+    placeholder,
+    validationHandler,
+    previousEnd = 0,
+    value = '',
+  } = props;
   const inputRef = useRef<HTMLInputElement | null>(null);
-  const [value, setValue] = useState('');
 
-  /**
-   * @description Resets input value to given
-   */
-  const resetValue = useCallback(() => {
-    // Todo: check if change is necessary
-    try {
-      setValue(stringFromMillis(time + delay));
-    } catch (error) {
-      emitError(`Unable to parse date: ${error}`);
-    }
-  }, [delay, emitError, time]);
-
-  /**
-   * @description Selects input text on focus
-   */
-  const handleFocus = useCallback(() => {
-    inputRef.current?.select();
-  }, []);
-
-  /**
-   * @description Submit handler
-   * @param {string} newValue
-   */
-  const handleSubmit = useCallback(
-    (newValue: string) => {
-      // Check if there is anything there
-      if (newValue === '') {
-        return false;
-      }
-
-      let newValMillis = 0;
-
-      // check for known aliases
-      if (newValue === 'p' || newValue === 'prev' || newValue === 'previous') {
-        // string to pass should be the time of the end before
-        if (previousEnd != null) {
-          newValMillis = previousEnd;
-        }
-      } else if (newValue.startsWith('+') || newValue.startsWith('p+') || newValue.startsWith('p +')) {
-        // string to pass should add to the end before
-        const val = newValue.substring(1);
-        newValMillis = previousEnd + forgivingStringToMillis(val);
-      } else {
-        // convert entered value to milliseconds
-        newValMillis = forgivingStringToMillis(newValue);
-      }
-
-      // Time now and time submittedVal
-      const originalMillis = time + delay;
-
-      // check if time is different from before
-      if (newValMillis === originalMillis) return false;
-
-      // validate with parent
-      if (!validationHandler(name, newValMillis)) return false;
-
-      // update entry
-      submitHandler(name, newValMillis);
-
-      return true;
-    },
-    [delay, name, previousEnd, submitHandler, time, validationHandler],
+  const timeInputProps = useTimeInput(
+    value,
+    inputRef,
+    time,
+    delay,
+    previousEnd,
+    name,
+    submitHandler,
+    validationHandler,
   );
-
-  /**
-   * @description Prepare time fields
-   * @param {string} value string to be parsed
-   */
-  const validateAndSubmit = useCallback(
-    (newValue: string) => {
-      const success = handleSubmit(newValue);
-      if (success) {
-        const ms = forgivingStringToMillis(newValue);
-        setValue(stringFromMillis(ms + delay));
-      } else {
-        resetValue();
-      }
-    },
-    [delay, handleSubmit, resetValue],
-  );
-
-  /**
-   * @description Handles common keys for submit and cancel
-   * @param {KeyboardEvent} event
-   */
-  const onKeyDownHandler = useCallback(
-    (event: KeyboardEvent<HTMLInputElement>) => {
-      if (event.key === 'Enter') {
-        inputRef.current?.blur();
-        validateAndSubmit((event.target as HTMLInputElement).value);
-      } else if (event.key === 'Tab') {
-        validateAndSubmit((event.target as HTMLInputElement).value);
-      }
-      if (event.key === 'Escape') {
-        inputRef.current?.blur();
-        resetValue();
-      }
-    },
-    [resetValue, validateAndSubmit],
-  );
-
-  useEffect(() => {
-    if (time == null) return;
-    resetValue();
-  }, [emitError, resetValue, time]);
 
   const isDelayed = delay != null && delay !== 0;
 
@@ -172,11 +84,7 @@ export default function TimeInput(props: TimeInputProps) {
         type='text'
         placeholder={placeholder}
         variant='ontime-filled'
-        onFocus={handleFocus}
-        onChange={(event) => setValue(event.target.value)}
-        onBlur={resetValue}
-        onKeyDown={onKeyDownHandler}
-        value={value}
+        {...timeInputProps}
         maxLength={8}
       />
     </InputGroup>
